% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Title}
\format{
A data frame with 3 variables:
\describe{
\item{Year}{Description of var1}
\item{GEO}{Description of var2}
\item{Funder}{Description of var3}
\item{Performer}{Description of var4}
\item{Science type}{Description of var5}
\item{Prices}{Description of var6}
\item{VALUE}{Description of var7}
}
}
\usage{
data
}
\description{
Title
}
\keyword{datasets}
