% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\docType{data}
\name{dd}
\alias{dd}
\title{Title}
\format{
A data frame with 3 variables:
\describe{
\item{YearMonth}{Description of var1}
\item{Year}{Description of var2}
\item{Month}{Description of var3}
\item{GEO}{Description of var3}
}
}
\usage{
dd
}
\description{
Title
}
\keyword{datasets}
